Staff badge launch fixes

+++ uber/site_sections/mits.py
@@ -272,7 +272,8 @@ class Root:
                     new_pic = add_new_image(pic, game)
                     session.add(new_pic)
                 if header_pic:
-                    session.delete(game.guidebook_header)
+                    if game.guidebook_header:
+                        session.delete(game.guidebook_header)
                     session.add(header_pic)
                 if thumbnail_pic:
                     session.delete(game.guidebook_thumbnail)
